// PRO-EP/pro-ep-ui/src/app/interfaces/interfaces.ts
export interface IUserMin {
  userid: number;
  dempoid: string;
  displayName: string | null;
  fname: string;
  lname: string;
  preferredfname: string;
  preferredlname: string;
  language: string;
  role: number;
  buddy?: boolean;
  employmenttype?: number;
  trainedon: string;
  trainedOnArray: string[];
  active?: boolean;
  canEdit?: boolean;
  selected?: boolean;
  checked?: boolean;
  changed?: boolean;
  schedulinglevel?: number;
}

export interface ITimeCard {
  timeCardId: number;
  dempoid: string;
  //workDate?: Date;
  timeIn?: Date | null;
  timeOut?: Date | null;
  entryBy: string;
  entryDt?: Date;
  modBy: string;
  modDt?: Date;
  datetimein?: Date | null;
  datetimeout?: Date | null;
  externalIp: string | null;
  //MachineName: string;
}
// ...
export interface ISchedule {
  preschedulekey: number;
  dempoid: string | null;
  fname: string | null;
  lname: string | null;
  preferredfname: string | null;
  preferredlname: string | null;
  displayName: string | null;
  userName: string | null;
  projectid?: number | null;
  projectName: string | null;
  projectColor: string;
  scheduledate: Date;
  comments: string | null;
  startdatetime?: Date | null;
  enddatetime?: Date | null;
  startTime?: string;
  endTime?: string;
  dayNumber?: number | null;
  weekNumber?: number | null;
  weekStart: Date | null;
  weekEnd: Date | null;
  dayOfWeek: number;
  dayOfWeekText?: string;
  monthNumber?: number | null;
  month: string | null;
  scheduleyear?: number | null;
  scheduledHours?: number | null;
  expr1: string | null;
  requestId?: number | null;
  requestDetails: string | null;
  requestCodeId?: number | null;
  requestCode: string | null;
  userid: number | null;
  trainedon: string | null;
  language: string | null;
  changed?: boolean;
  saved?: boolean | null;
  markedForDeletion?: boolean | null;
  addTab?: boolean;
  addInitial?: boolean;
  invalid?: boolean;
  invalidFields?: string[];
  entryBy: string | null;
  entryDt?: Date;
  validationMessages?: string[];
  hoverMessage?: string;
  scheduleConflict?: boolean;
  justAdded?: boolean;
  initialProjectid?: number | null;
}
// ...
export interface IWeekStartAndEnd {
  weekStart: Date;
  weekEnd: Date;
}
// ...
export interface IAuthenticatedUser {
  displayName: string;
  duDukeID: string;
  eppn: string;
  netID: string;
  isMemberOf: string;
  interviewer: boolean;
  resourceGroup: boolean;
  admin: boolean;
  userRoles?: string[] | undefined;
  sessionMinsLeft: number;
  timecards: ITimeCard[];
}
// ...
export interface IProjectTotals {
  projectid?: number;
  projectName: string;
  scheduledate: string;
  dayNumber?: number;
  weekNumber?: number;
  yearNumber?: number;
  startdatetime: Date;
  enddatetime: Date;
  totalHours?: number;
}

export interface IProjectTotalsSummed {
  total?: number;
  yearnumber?: number;
  weeknumber?: number;
  daynumber?: number;
  scheduledate: string;
  startdatetime: Date;
}

export interface IProjectTotalsReport {
  projectTotals: IProjectTotals[];
  projectTotalsSummed: IProjectTotalsSummed[];
}

export interface IProjectTotalsReportFlat {
  projectColumn: string;
  column1: number;
  column2: number;
  column3: number;
  column4: number;
  column5: number;
  column6: number;
  column7: number;
  totalColumn: number;
}